pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proposal-bot"
version = "1.0.0"
description = "LangChain Deep Agent system for automated market research proposal generation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Team", email = "team@example.com"}
]
keywords = ["langchain", "deep-agents", "langgraph", "market-research", "proposals"]

dependencies = [
    # LangChain Core
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",

    # Google Integrations
    "langchain-google-community[gmail]>=2.0.0",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.120.0",

    # Anthropic
    "anthropic>=0.34.0",

    # Database & Storage
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.0",
    "pgvector>=0.3.0",

    # Web Framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",

    # Monitoring & Logging
    "structlog>=24.1.0",
    "sentry-sdk[fastapi]>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[tool.setuptools]
packages = ["proposal_bot"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
